+++ src/main/java/io/leangen/graphql/generator/mapping/ConverterRegistry.java
@@ -1,10 +1,13 @@
 package io.leangen.graphql.generator.mapping;
 
+import io.leangen.graphql.metadata.Operation;
+import io.leangen.graphql.metadata.Resolver;
 import io.leangen.graphql.util.ClassUtils;
 
 import java.lang.reflect.AnnotatedType;
 import java.util.Collections;
 import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * @author Bojan Tomic (kaqqao)
@@ -31,6 +34,17 @@ public class ConverterRegistry {
     @SuppressWarnings("unchecked")
     public <T, S> OutputConverter<T, S> getOutputConverter(AnnotatedType outputType) {
         return (OutputConverter<T, S>) outputConverters.stream().filter(conv -> conv.supports(outputType)).findFirst().orElse(null);
+    }
+
+    public ConverterRegistry forOperation(Operation operation) {
+        List<AnnotatedType> returnTypes = operation.getResolvers().stream().map(Resolver::getReturnType).collect(Collectors.toList());
+        List<OutputConverter> filteredByOperation = outputConverters.stream()
+                .filter(conv -> conv.supports(new ConverterSupportParams(operation.getOperationType())))
+                .collect(Collectors.toList());
+        if (filteredByOperation.stream().noneMatch(conv -> returnTypes.stream().anyMatch(conv::supports))) {
+            filteredByOperation = Collections.emptyList();
+        }
+        return new ConverterRegistry(inputConverters, filteredByOperation);
     }
 
     public AnnotatedType getMappableInputType(AnnotatedType type) {