+++ src/main/java/io/leangen/graphql/generator/mapping/OutputConverter.java
@@ -1,8 +1,8 @@
 package io.leangen.graphql.generator.mapping;
 
-import java.lang.reflect.AnnotatedType;
-
 import io.leangen.graphql.execution.ResolutionEnvironment;
+
+import java.lang.reflect.AnnotatedType;
 
 /**
  * @author Bojan Tomic (kaqqao)
@@ -12,4 +12,8 @@ public interface OutputConverter<T, S> {
     S convertOutput(T original, AnnotatedType type, ResolutionEnvironment resolutionEnvironment);
     
     boolean supports(AnnotatedType type);
+
+    default boolean supports(ConverterSupportParams params) {
+        return true;
+    }
 }