+++ src/main/java/io/leangen/graphql/metadata/strategy/query/DefaultOperationBuilder.java
@@ -1,5 +1,6 @@
 package io.leangen.graphql.metadata.strategy.query;
 
+import graphql.language.OperationDefinition;
 import io.leangen.geantyref.GenericTypeReflector;
 import io.leangen.graphql.annotations.GraphQLUnion;
 import io.leangen.graphql.execution.GlobalEnvironment;
@@ -57,21 +58,25 @@ public class DefaultOperationBuilder implements OperationBuilder {
 
     @Override
     public Operation buildQuery(Type contextType, List<Resolver> resolvers, GlobalEnvironment environment) {
-        String name = resolveName(resolvers);
-        AnnotatedType javaType = resolveJavaType(name, resolvers, environment.messageBundle);
-        List<OperationArgument> arguments = collectArguments(name, resolvers);
-        boolean batched = isBatched(resolvers);
-        return new Operation(name, javaType, contextType, arguments, resolvers, batched);
+        return buildOperation(contextType, resolvers, OperationDefinition.Operation.QUERY, environment);
     }
 
     @Override
     public Operation buildMutation(Type context, List<Resolver> resolvers, GlobalEnvironment environment) {
-        return buildQuery(context, resolvers, environment);
+        return buildOperation(context, resolvers, OperationDefinition.Operation.MUTATION, environment);
     }
 
     @Override
     public Operation buildSubscription(Type context, List<Resolver> resolvers, GlobalEnvironment environment) {
-        return buildQuery(context, resolvers, environment);
+        return buildOperation(context, resolvers, OperationDefinition.Operation.SUBSCRIPTION, environment);
+    }
+
+    private Operation buildOperation(Type contextType, List<Resolver> resolvers, OperationDefinition.Operation operationType, GlobalEnvironment environment) {
+        String name = resolveName(resolvers);
+        AnnotatedType javaType = resolveJavaType(name, resolvers, environment.messageBundle);
+        List<OperationArgument> arguments = collectArguments(name, resolvers);
+        boolean batched = isBatched(resolvers);
+        return new Operation(name, javaType, contextType, arguments, resolvers, operationType, batched);
     }
 
     protected String resolveName(List<Resolver> resolvers) {